Add try await keywords when chopping veggies
Is this code example missing the `try await` keywords when chopping veggies? In the previous example, inside the loop that would not produce any meaningful concurrency, line 190 is written:

```swift
veggies[i] = try await veggies[i].chopped()
```

It seems like the `chopped()` method  has the potential for a kitchen knife incident for each veggie. I assume this needs to be accounted for in the meaningful concurrency example too?